feat: split script UseCase by actions, minor changes
