+++ src/data/user-monitoring/entities/MSTeamsWebhookOptions.ts
@@ -1,4 +1,4 @@
-export interface WebhookOptions {
+export interface MSTeamsWebhookOptions {
     ms_url: string;
     proxy: string;
     server_name: string;
